refactor: extract daemon setup and transport wiring into AppDelegate+DaemonSetup.swift
